requirements.txt
requests
PyYAML
sounddevice
numpy
Flask
Flask-SocketIO
shazamio